Clean BackendService
